+++ src/Storefront/Framework/Cookie/PayPalCookieProvider.php
@@ -7,35 +7,28 @@
 
 namespace Swag\PayPal\Storefront\Framework\Cookie;
 
-use Shopware\Core\Framework\Context;
-use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
-use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
-use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
-use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\PrefixFilter;
 use Shopware\Core\Framework\Log\Package;
-use Shopware\Core\Framework\Uuid\Uuid;
+use Shopware\Core\PlatformRequest;
 use Shopware\Storefront\Framework\Cookie\CookieProviderInterface;
+use Swag\PayPal\Util\PaymentMethodUtil;
 use Symfony\Component\HttpFoundation\RequestStack;
 
 #[Package('checkout')]
 class PayPalCookieProvider implements CookieProviderInterface
 {
-    private CookieProviderInterface $original;
-
     /**
      * @internal
      */
     public function __construct(
-        CookieProviderInterface $cookieProvider,
-        private readonly EntityRepository $paymentMethodRepository,
+        private readonly CookieProviderInterface $cookieProvider,
+        private readonly PaymentMethodUtil $paymentMethodUtil,
         private readonly RequestStack $requestStack,
     ) {
-        $this->original = $cookieProvider;
     }
 
     public function getCookieGroups(): array
     {
-        $cookies = $this->original->getCookieGroups();
+        $cookies = $this->cookieProvider->getCookieGroups();
 
         foreach ($cookies as &$cookie) {
             if (!\is_array($cookie)) {
@@ -71,19 +64,11 @@ class PayPalCookieProvider implements CookieProviderInterface
 
     private function isPayPalPaymentActive(): bool
     {
-        $criteria = new Criteria();
-        $criteria->setLimit(1);
-        $criteria->addFilter(new EqualsFilter('active', true));
-        $criteria->addFilter(new PrefixFilter('technicalName', 'swag_paypal_'));
-
-        $mainRequestAttributes = $this->requestStack->getMainRequest()?->attributes;
-        $context = $mainRequestAttributes?->get('sw-context') ?? Context::createDefaultContext();
-        $salesChannelId = $mainRequestAttributes?->getString('sw-sales-channel-id') ?? '';
-
-        if (Uuid::isValid($salesChannelId)) {
-            $criteria->addFilter(new EqualsFilter('salesChannels.id', $salesChannelId));
+        $salesChannelContext = $this->requestStack->getMainRequest()?->attributes->get(PlatformRequest::ATTRIBUTE_SALES_CHANNEL_CONTEXT_OBJECT);
+        if (!$salesChannelContext) {
+            return false;
         }
 
-        return $this->paymentMethodRepository->searchIds($criteria, $context)->firstId() !== null;
+        return $this->paymentMethodUtil->isPaymentMethodActive($salesChannelContext);
     }
 }