+++ tests/Installment/Banner/InstallmentBannerSubscriberTest.php
@@ -101,7 +101,7 @@ class InstallmentBannerSubscriberTest extends TestCase
     public function testAddInstallmentBannerPayPalNotInSalesChannel(): void
     {
         $event = $this->createCheckoutCartPageLoadedEvent(false);
-        $this->paymentMethodUtil->expects(static::once())->method('isPaypalPaymentMethodInSalesChannel')->willReturn(false);
+        $this->paymentMethodUtil->expects(static::once())->method('isPaymentMethodActive')->willReturn(false);
 
         $this->createInstallmentBannerSubscriber()->addInstallmentBanner($event);
 
@@ -111,7 +111,7 @@ class InstallmentBannerSubscriberTest extends TestCase
     public function testAddInstallmentBannerInvalidSettings(): void
     {
         $event = $this->createCheckoutCartPageLoadedEvent();
-        $this->paymentMethodUtil->expects(static::once())->method('isPaypalPaymentMethodInSalesChannel')->willReturn(true);
+        $this->paymentMethodUtil->expects(static::once())->method('isPaymentMethodActive')->willReturn(true);
 
         $this->createInstallmentBannerSubscriber([
             Settings::CLIENT_ID => null,
@@ -124,7 +124,7 @@ class InstallmentBannerSubscriberTest extends TestCase
     public function testAddInstallmentBannerCheckoutCartDisabled(): void
     {
         $event = $this->createCheckoutCartPageLoadedEvent();
-        $this->paymentMethodUtil->expects(static::once())->method('isPaypalPaymentMethodInSalesChannel')->willReturn(true);
+        $this->paymentMethodUtil->expects(static::once())->method('isPaymentMethodActive')->willReturn(true);
 
         $this->createInstallmentBannerSubscriber([
             Settings::INSTALLMENT_BANNER_CART_ENABLED => false,
@@ -139,7 +139,7 @@ class InstallmentBannerSubscriberTest extends TestCase
     public function testAddInstallmentBannerCheckoutCart(): void
     {
         $event = $this->createCheckoutCartPageLoadedEvent();
-        $this->paymentMethodUtil->expects(static::once())->method('isPaypalPaymentMethodInSalesChannel')->willReturn(true);
+        $this->paymentMethodUtil->expects(static::once())->method('isPaymentMethodActive')->willReturn(true);
 
         $this->createInstallmentBannerSubscriber()->addInstallmentBanner($event);
 
@@ -162,7 +162,7 @@ class InstallmentBannerSubscriberTest extends TestCase
     public function testAddInstallmentBannerCheckoutCartExcludedProduct(): void
     {
         $event = $this->createCheckoutCartPageLoadedEvent();
-        $this->paymentMethodUtil->expects(static::once())->method('isPaypalPaymentMethodInSalesChannel')->willReturn(true);
+        $this->paymentMethodUtil->expects(static::once())->method('isPaymentMethodActive')->willReturn(true);
         $this->excludedProductValidator->expects(static::once())->method('cartContainsExcludedProduct')->willReturn(true);
 
         $this->createInstallmentBannerSubscriber()->addInstallmentBanner($event);
@@ -173,7 +173,7 @@ class InstallmentBannerSubscriberTest extends TestCase
     public function testAddInstallmentBannerOffCanvasCartDisabled(): void
     {
         $event = $this->createOffCanvasCartPageLoadedEvent();
-        $this->paymentMethodUtil->expects(static::once())->method('isPaypalPaymentMethodInSalesChannel')->willReturn(true);
+        $this->paymentMethodUtil->expects(static::once())->method('isPaymentMethodActive')->willReturn(true);
 
         $this->createInstallmentBannerSubscriber([
             Settings::INSTALLMENT_BANNER_OFF_CANVAS_CART_ENABLED => false,
@@ -188,7 +188,7 @@ class InstallmentBannerSubscriberTest extends TestCase
     public function testAddInstallmentBannerProductPage(): void
     {
         $event = $this->createProductPageLoadedEvent();
-        $this->paymentMethodUtil->expects(static::once())->method('isPaypalPaymentMethodInSalesChannel')->willReturn(true);
+        $this->paymentMethodUtil->expects(static::once())->method('isPaymentMethodActive')->willReturn(true);
 
         $this->createInstallmentBannerSubscriber()->addInstallmentBanner($event);
 
@@ -202,7 +202,7 @@ class InstallmentBannerSubscriberTest extends TestCase
     public function testAddInstallmentBannerProductPageDisabled(): void
     {
         $event = $this->createProductPageLoadedEvent();
-        $this->paymentMethodUtil->expects(static::once())->method('isPaypalPaymentMethodInSalesChannel')->willReturn(true);
+        $this->paymentMethodUtil->expects(static::once())->method('isPaymentMethodActive')->willReturn(true);
 
         $this->createInstallmentBannerSubscriber([
             Settings::INSTALLMENT_BANNER_DETAIL_PAGE_ENABLED => false,
@@ -217,7 +217,7 @@ class InstallmentBannerSubscriberTest extends TestCase
     public function testAddInstallmentBannerProductPageExcludedProduct(): void
     {
         $event = $this->createProductPageLoadedEvent();
-        $this->paymentMethodUtil->expects(static::once())->method('isPaypalPaymentMethodInSalesChannel')->willReturn(true);
+        $this->paymentMethodUtil->expects(static::once())->method('isPaymentMethodActive')->willReturn(true);
         $this->excludedProductValidator->expects(static::once())->method('isProductExcluded')->willReturn(true);
 
         $this->createInstallmentBannerSubscriber()->addInstallmentBanner($event);
@@ -228,7 +228,7 @@ class InstallmentBannerSubscriberTest extends TestCase
     public function testAddInstallmentBannerProductPageWithAdvancedPrices(): void
     {
         $event = $this->createProductPageLoadedEvent(true);
-        $this->paymentMethodUtil->expects(static::once())->method('isPaypalPaymentMethodInSalesChannel')->willReturn(true);
+        $this->paymentMethodUtil->expects(static::once())->method('isPaymentMethodActive')->willReturn(true);
 
         $this->createInstallmentBannerSubscriber()->addInstallmentBanner($event);
 
@@ -242,7 +242,7 @@ class InstallmentBannerSubscriberTest extends TestCase
     public function testAddInstallmentBannerFooterPayPalNotInSalesChannel(): void
     {
         $event = $this->createFooterPageletLoadedEvent(false);
-        $this->paymentMethodUtil->expects(static::once())->method('isPaypalPaymentMethodInSalesChannel')->willReturn(false);
+        $this->paymentMethodUtil->expects(static::once())->method('isPaymentMethodActive')->willReturn(false);
 
         $this->createInstallmentBannerSubscriber()->addInstallmentBannerPagelet($event);
 
@@ -252,7 +252,7 @@ class InstallmentBannerSubscriberTest extends TestCase
     public function testAddInstallmentBannerFooterInvalidSettings(): void
     {
         $event = $this->createFooterPageletLoadedEvent();
-        $this->paymentMethodUtil->expects(static::once())->method('isPaypalPaymentMethodInSalesChannel')->willReturn(true);
+        $this->paymentMethodUtil->expects(static::once())->method('isPaymentMethodActive')->willReturn(true);
 
         $this->createInstallmentBannerSubscriber([
             Settings::CLIENT_ID => null,
@@ -265,7 +265,7 @@ class InstallmentBannerSubscriberTest extends TestCase
     public function testAddInstallmentBannerFooterDisabled(): void
     {
         $event = $this->createFooterPageletLoadedEvent();
-        $this->paymentMethodUtil->expects(static::once())->method('isPaypalPaymentMethodInSalesChannel')->willReturn(true);
+        $this->paymentMethodUtil->expects(static::once())->method('isPaymentMethodActive')->willReturn(true);
 
         $this->createInstallmentBannerSubscriber([
             Settings::INSTALLMENT_BANNER_FOOTER_ENABLED => false,
@@ -280,7 +280,7 @@ class InstallmentBannerSubscriberTest extends TestCase
     public function testAddInstallmentBannerFooterPagelet(): void
     {
         $event = $this->createFooterPageletLoadedEvent();
-        $this->paymentMethodUtil->expects(static::once())->method('isPaypalPaymentMethodInSalesChannel')->willReturn(true);
+        $this->paymentMethodUtil->expects(static::once())->method('isPaymentMethodActive')->willReturn(true);
 
         $this->createInstallmentBannerSubscriber()->addInstallmentBannerPagelet($event);
 
@@ -294,7 +294,7 @@ class InstallmentBannerSubscriberTest extends TestCase
     public function testAddInstallmentBannerAccountLoginPageDisabled(): void
     {
         $event = $this->createCheckoutRegisterPageLoadedEvent();
-        $this->paymentMethodUtil->expects(static::once())->method('isPaypalPaymentMethodInSalesChannel')->willReturn(true);
+        $this->paymentMethodUtil->expects(static::once())->method('isPaymentMethodActive')->willReturn(true);
 
         $this->createInstallmentBannerSubscriber([
             Settings::INSTALLMENT_BANNER_LOGIN_PAGE_ENABLED => false,