+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 8.9.1
+- Fixes an issue, where required cookies were not displayed in the banner even though PayPal scripts had been loaded (shopware/SwagPayPal#506)
+- Fixes an issue, where languages not supported by PayPal did not fall back to a supported language (shopware/shopware#13950)
+
 # 8.9.0
 - Fixes an issue, where the express checkout could choose a customer country that was not assigned to the correct sales channel (shopware/SwagPayPal#479)
 - Fixes an issue, where cookies are added even though associated payment methods are not active (shopware/SwagPayPal#457)